requirements.txt
langchain==0.0.352
openai==1.3.7
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0
streamlit==1.29.0
python-dotenv==1.0.0
pytz==2023.3.post1
email-validator==2.1.0.post1
requests==2.31.0
pandas==2.1.4